% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haawe.R
\name{haawe}
\alias{haawe}
\title{Prepares spacial data files to be loaded}
\usage{
haawe(x, keyname = NULL)
}
\arguments{
\item{x}{A keyword for the data to be loaded or alternatively the URL where the data is located}

\item{keyname}{The keyword to be assigned to the data in case that x is a URL}
}
\value{
Generated R script saved in the data directory of the kokua package.
}
\description{
\code{haawe} Organizes loaded data and generates an R script to load the data.
}
\details{
stub
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}
